requirements.txt
fastapi
uvicorn
snowflake-connector-python
streamlit
requests
python-dotenv
streamlit
pandas